+++ custom_components/bms_ble/manifest.json
@@ -43,7 +43,7 @@
   "integration_type": "device",
   "iot_class": "local_polling",
   "issue_tracker": "https://github.com/patman15/BMS_BLE-HA/issues",
-  "loggers": ["bms_ble", "daly_bms", "jbd_bms", "jikong_bms", "ogt_bms", "seplos_bms"],
+  "loggers": ["bms_ble", "daly_bms", "dpwercore_bms", "jbd_bms", "jikong_bms", "ogt_bms", "seplos_bms"],
   "requirements": [],
-  "version": "1.7.1"
+  "version": "1.8.0"
 }
