+++ custom_components/bms_ble/plugins/dpwrcore_bms.py
